add file delete, toasts

+++ server/src/routes/file.ts
@@ -1,6 +1,9 @@
 import { z } from "zod";
 import { router, publicProcedure } from "../trpc";
 import { PrismaClient } from "@prisma/client";
+import fs from "fs";
+import { UPLOADS_DIR } from "..";
+import path from "path";
 
 export const prisma = new PrismaClient();
 
@@ -12,12 +15,26 @@ export const fileRouter = router({
     return files.reverse();
   }),
   deleteById: publicProcedure.input(z.number()).mutation(async ({ input }) => {
+    const toDelete = await prisma.file.findUnique({ where: { id: input } });
+    // TODO: Custom error message
+    if (!toDelete) return null;
+
+    if (fs.existsSync(path.join(UPLOADS_DIR, toDelete.filename)))
+      fs.unlinkSync(path.join(UPLOADS_DIR, toDelete.filename));
+
     const deleted = await prisma.file.delete({ where: { id: input } });
 
     return deleted;
   }),
   deleteAll: publicProcedure.mutation(async (req) => {
-    await prisma.file.deleteMany({});
+    const allFiles = await prisma.file.findMany();
+
+    for (const file of allFiles) {
+      if (fs.existsSync(path.join(UPLOADS_DIR, file.filename)))
+        fs.unlinkSync(path.join(UPLOADS_DIR, file.filename));
+
+      await prisma.file.delete({ where: { id: file.id } });
+    }
 
     return;
   }),
